Massive number of files changed to accomodate new requirement of JANA 0.7.4 that event number be uint64_t. Most of these changes were done using search and replace with these commands:
find . -type f -name "*.h" -exec sed -i "" 's/int\ runnumber/int32_t\ runnumber/g' {} \;
find . -type f -name "*.cc" -exec sed -i "" 's/int\ runnumber/int32_t\ runnumber/g' {} \;
find . -type f -name "*.h" -exec sed -i "" 's/int\ eventnumber/uint64_t\ eventnumber/g' {} \;
find . -type f -name "*.cc" -exec sed -i "" 's/int\ eventnumber/uint64_t\ eventnumber/g' {} \;
find . -type f -name "*.cc" -exec sed -i "" 's/unsigned\ int32_t\ runnumber/int32_t\ runnumber/g' {} \;
find . -type f -name "*.h" -exec sed -i "" 's/unsigned\ int32_t\ runnumber/int32_t\ runnumber/g' {} \;
find . -type f -name "*.cc" -exec sed -i "" 's/int\ eventNo/uint64_t\ eventNo/g' {} \;
find . -type f -name "*.h" -exec sed -i "" 's/int\ eventNo/uint64_t\ eventNo/g' {} \;
find . -type f -name "*.h" -exec sed -i "" 's/int\ locEventNumber/uint64_t\ locEventNumber/g' {} \;
find . -type f -name "*.cc" -exec sed -i "" 's/int\ locEventNumber/uint64_t\ locEventNumber/g' {} \;

+++ src/libraries/ANALYSIS/DAnalysisUtilities_factory.h
@@ -18,7 +18,7 @@ class DAnalysisUtilities_factory : public jana::JFactory<DAnalysisUtilities>
 		~DAnalysisUtilities_factory(){};
   
 	private:
-		jerror_t evnt(jana::JEventLoop *loop, int eventnumber)
+		jerror_t evnt(jana::JEventLoop *loop, uint64_t eventnumber)
 		{
 			// Create single DAnalysisUtilities object and mark the factory as
 			// persistent so it doesn't get deleted every event.
