+++ src/libraries/FDC/DFDCPseudo_factory.cc
@@ -114,7 +114,7 @@ jerror_t DFDCPseudo_factory::init(void)
 //------------------
 // brun
 //------------------
-jerror_t DFDCPseudo_factory::brun(JEventLoop *loop, int runnumber)
+jerror_t DFDCPseudo_factory::brun(JEventLoop *loop, int32_t runnumber)
 {
   // Get pointer to DGeometry object
   DApplication* dapp=dynamic_cast<DApplication*>(eventLoop->GetJApplication());
@@ -232,7 +232,7 @@ jerror_t DFDCPseudo_factory::erun(void){
 /// DFDCPseudo_factory::evnt():
 /// this is the place that anode hits and DFDCCathodeClusters are organized into pseudopoints.
 ///
-jerror_t DFDCPseudo_factory::evnt(JEventLoop* eventLoop, int eventNo) {
+jerror_t DFDCPseudo_factory::evnt(JEventLoop* eventLoop, uint64_t eventNo) {
   if (!USE_FDC) return NOERROR;
 
 	// Get all FDC hits (anode and cathode)
