+++ src/libraries/TAGGER/DTAGMHit_factory.cc
@@ -56,7 +56,7 @@ jerror_t DTAGMHit_factory::init(void)
 //------------------
 // brun
 //------------------
-jerror_t DTAGMHit_factory::brun(jana::JEventLoop *eventLoop, int runnumber)
+jerror_t DTAGMHit_factory::brun(jana::JEventLoop *eventLoop, int32_t runnumber)
 {
     // Only print messages for one thread whenever run number change
     static pthread_mutex_t print_mutex = PTHREAD_MUTEX_INITIALIZER;
@@ -104,7 +104,7 @@ jerror_t DTAGMHit_factory::brun(jana::JEventLoop *eventLoop, int runnumber)
 //------------------
 // evnt
 //------------------
-jerror_t DTAGMHit_factory::evnt(JEventLoop *loop, int eventnumber)
+jerror_t DTAGMHit_factory::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
     /// Generate DTAGMHit object for each DTAGMDigiHit object.
     /// This is where the first set of calibration constants
