+++ src/libraries/TRACKING/DTrackTimeBased_factory.h
@@ -37,8 +37,8 @@ class DTrackTimeBased_factory:public jana::JFactory<DTrackTimeBased>{
   
  private:
   jerror_t init(void);						///< Called once at program start.
-  jerror_t brun(jana::JEventLoop *loop, int runnumber);	///< Called everytime a new run number is detected.
-  jerror_t evnt(jana::JEventLoop *loop, int eventnumber);	///< Called every event.
+  jerror_t brun(jana::JEventLoop *loop, int32_t runnumber);	///< Called everytime a new run number is detected.
+  jerror_t evnt(jana::JEventLoop *loop, uint64_t eventnumber);	///< Called every event.
   jerror_t erun(void);						///< Called everytime run number changes, provided brun has been called.
   jerror_t fini(void);						///< Called after last event of last event source has been processed.
