+++ src/plugins/Analysis/DAQTreeBCAL/JEventProcessor_DAQTreeBCAL.cc
@@ -94,7 +94,7 @@ jerror_t JEventProcessor_DAQTreeBCAL::init(void)
 //------------------
 // brun
 //------------------
-jerror_t JEventProcessor_DAQTreeBCAL::brun(JEventLoop *eventLoop, int runnumber)
+jerror_t JEventProcessor_DAQTreeBCAL::brun(JEventLoop *eventLoop, int32_t runnumber)
 {
 	// This is called whenever the run number changes
 	return NOERROR;
@@ -103,7 +103,7 @@ jerror_t JEventProcessor_DAQTreeBCAL::brun(JEventLoop *eventLoop, int runnumber)
 //------------------
 // evnt
 //------------------
-jerror_t JEventProcessor_DAQTreeBCAL::evnt(JEventLoop *loop, int eventnumber)
+jerror_t JEventProcessor_DAQTreeBCAL::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
 	eventnum = eventnumber;
